Fixed #1427: Expanded the list of invalid ampersand locations.
Also moved magic contents to Magic.
